refactor: 아이템에 있는 AllergyType, Converter로 통합 (#14)
