+++ src/main/java/com/devlop/siren/domain/user/domain/User.java
@@ -1,6 +1,7 @@
 package com.devlop.siren.domain.user.domain;
 
-import com.devlop.siren.domain.user.util.AllergyConverter;
+import com.devlop.siren.domain.item.entity.AllergyType;
+import com.devlop.siren.domain.item.utils.AllergyConverter;
 import com.devlop.siren.domain.user.util.validator.KoreanNickname;
 import com.devlop.siren.global.common.BaseEntity;
 import lombok.AccessLevel;
