+++ src/main/java/com/devlop/siren/domain/user/dto/request/UserRegisterRequest.java
@@ -1,6 +1,6 @@
 package com.devlop.siren.domain.user.dto.request;
 
-import com.devlop.siren.domain.user.domain.AllergyType;
+import com.devlop.siren.domain.item.entity.AllergyType;
 import com.devlop.siren.domain.user.domain.User;
 import com.devlop.siren.domain.user.domain.UserRole;
 import com.devlop.siren.domain.user.util.validator.KoreanNickname;
