+++ src/test/java/com/devlop/siren/user/repository/UserRepositoryTest.java
@@ -1,8 +1,8 @@
 package com.devlop.siren.user.repository;
 
+import com.devlop.siren.domain.item.entity.AllergyType;
 import com.devlop.siren.fixture.UserFixture;
 import com.devlop.siren.domain.user.repository.UserRepository;
-import com.devlop.siren.domain.user.domain.AllergyType;
 import com.devlop.siren.domain.user.domain.User;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
