+++ src/test/java/com/devlop/siren/user/service/UserServiceTest.java
@@ -1,6 +1,7 @@
 package com.devlop.siren.user.service;
 
-import com.devlop.siren.domain.user.domain.AllergyType;
+import com.devlop.siren.domain.item.entity.AllergyType;
+import com.devlop.siren.domain.item.utils.AllergyConverter;
 import com.devlop.siren.domain.user.domain.User;
 import com.devlop.siren.domain.user.domain.UserRole;
 import com.devlop.siren.domain.user.dto.UserTokenDto;
@@ -9,7 +10,6 @@ import com.devlop.siren.domain.user.dto.request.UserRegisterRequest;
 import com.devlop.siren.domain.user.repository.UserRepository;
 import com.devlop.siren.domain.user.service.RedisService;
 import com.devlop.siren.domain.user.service.UserService;
-import com.devlop.siren.domain.user.util.AllergyConverter;
 import com.devlop.siren.fixture.UserFixture;
 import com.devlop.siren.global.exception.GlobalException;
 import com.devlop.siren.global.util.JwtTokenUtils;
